Major refactoring
Separated files into two modules: data_sources and graphql_types

The `data_sources` module contains logic solely responsible for
retrieving data from the various sources that the api has access to
(postgres, redis, json apis, etc) and defines types that represent data
stored in these sources.

All code related to defining the graphql schema has been moved to the 
graphql_types module. Types defined here represent data as it will be
presented to the user. Types defined in this module are responsible for 
building themselves from the types in the data_sources module, typically
by implementing the From trait for one or more of these types.

As part of this refactor, the graphql framework was switched from 
[juniper](https://docs.rs/juniper/0.12.0/juniper/) to 
[async_graphql](https://docs.rs/async-graphql/2.8.2/async_graphql/),
which is very similar, but apparently more stable and with some nicer
features.