more refactoring of schedule file
Schedule is now its own type, and contains the resolver for fetching its
list of timeslots, as well as extra metadata.

Seasons are no longer kept as strings, but instead as their own Season
enum, which can be either one of the most common types: SPRING, FALL,
SUMMER, or Other, which can represent any other string. The season type
seamlessly converts to and from Strings, and its purpose is mainly to
help rust optimize the code and to make the Schedule type more explicit.

Began adding doc comments